--- app/src/main/java/com/kuit/kuit5/HomeScreen.kt
@@ -1,13 +1,0 @@
-package com.kuit.kuit5
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun HomeScreen(modifier: Modifier = Modifier) {
-    Column(modifier = modifier) {
-        Text(text = "Home Screen")
-    }
-}
