+++ app/src/main/java/com/kuit/kuit5/navigation/KuitNavGraph.kt
@@ -5,11 +5,11 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.kuit.kuit5.AssetsScreen
-import com.kuit.kuit5.HealthScreen
-import com.kuit.kuit5.HomeScreen
-import com.kuit.kuit5.RecordsScreen
-import com.kuit.kuit5.ShoppingScreen
+import com.kuit.kuit5.ui.assets.screen.AssetsScreen
+import com.kuit.kuit5.ui.health.screen.HealthScreen
+import com.kuit.kuit5.ui.home.screen.HomeScreen
+import com.kuit.kuit5.ui.record.screen.RecordsScreen
+import com.kuit.kuit5.ui.shopping.screen.ShoppingScreen
 
 @Composable
 fun KuitNavGraph(
