+++ app/src/main/java/com/kuit/kuit5/ui/assets/screen/AssetsScreen.kt
@@ -1,0 +1,45 @@
+package com.kuit.kuit5.ui.assets.screen
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AssetsScreen(modifier: Modifier = Modifier) {
+    Column(modifier = modifier
+        .fillMaxSize()
+        .background(Color.White)
+    ) {
+    TopAppBar(
+        modifier = Modifier.fillMaxWidth(),
+        colors = TopAppBarDefaults.topAppBarColors(
+            containerColor = Color.White
+        ),
+        title = { Text("자산") },
+        actions = {
+            Row {
+                
+            }
+        }
+    )
+
+    }
+}
+
+@Preview
+@Composable
+private fun AssetsScreenPreview() {
+    AssetsScreen()
+
+}
