+++ app/src/main/java/com/kuit/kuit5/ui/home/screen/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.kuit.kuit5
+package com.kuit.kuit5.ui.home.screen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
@@ -6,8 +6,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
 @Composable
-fun AssetsScreen(modifier: Modifier = Modifier) {
+fun HomeScreen(modifier: Modifier = Modifier) {
     Column(modifier = modifier) {
-        Text(text = "Assets Screen")
+        Text(text = "Home Screen")
     }
 }
