+++ app/src/main/java/com/kuit/kuit5/ui/shopping/screen/ShoppingScreen.kt
@@ -1,4 +1,4 @@
-package com.kuit.kuit5
+package com.kuit.kuit5.ui.shopping.screen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
@@ -6,8 +6,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
 @Composable
-fun RecordsScreen(modifier: Modifier = Modifier) {
+fun ShoppingScreen(modifier: Modifier = Modifier) {
     Column(modifier = modifier) {
-        Text(text = "Records Screen")
+        Text(text = "Shopping Screen")
     }
 }
